Update Circle CI site image
Looks like the image we were using is no longer supported. Changing to rely on latest image instead. This may not be the best approach either, because new images might not support our needs, but pointing to a specific tag is more sure to break down the road than pointing to the latest one. Unless we want to point to a tag we control, which, may not be the worst idea.